requirements.txt
# OmniQuant - Unified Quantitative Trading Framework
# Python Dependencies

# ============================================================================
# Core Data Processing
# ============================================================================
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
pywavelets>=1.4.1
statsmodels>=0.14.0
pyarrow>=13.0.0

# ============================================================================
# Machine Learning & Deep Learning
# ============================================================================
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.0
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0  # Optional, for some models
transformers>=4.30.0
# statsmodels - already listed above

# ============================================================================
# Reinforcement Learning
# ============================================================================
stable-baselines3>=2.1.0
gymnasium>=0.29.0
ray[rllib]>=2.7.0
tensorboard>=2.13.0

# ============================================================================
# Causal Inference
# ============================================================================
dowhy>=0.11.0
econml>=0.14.0
pgmpy>=0.1.23
networkx>=3.1

# ============================================================================
# Portfolio Optimization
# ============================================================================
cvxpy>=1.4.0
pymoo>=0.6.0
optuna>=3.3.0
PyPortfolioOpt>=1.5.5

# ============================================================================
# Financial Data & Analysis
# ============================================================================
yfinance>=0.2.28
pandas-ta>=0.3.14b0
# ta-lib>=0.4.28  # Requires system TA-Lib C library (apt-get install libta-lib-dev)
arch>=6.2.0  # GARCH models

# ============================================================================
# Time Series
# ============================================================================
prophet>=1.1.4
pmdarima>=2.0.3
sktime>=0.24.0

# ============================================================================
# Visualization
# ============================================================================
plotly>=5.17.0
dash>=2.14.0
streamlit>=1.27.0
bokeh>=3.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# ============================================================================
# Database & Storage
# ============================================================================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
clickhouse-driver>=0.2.6
# redis - already listed below
pymongo>=4.5.0
duckdb>=0.9.0
arcticdb>=4.0.0

# ============================================================================
# Messaging & Serialization (Phase 2.0)
# ============================================================================
msgpack>=1.0.7
confluent-kafka>=2.3.0  # Kafka / Redpanda client
# aeron-python - requires Aeron C media driver (optional)

# ============================================================================
# Dependency Injection & CLI
# ============================================================================
dependency-injector>=4.41.0
typer>=0.9.0

# ============================================================================
# Distributed Computing
# ============================================================================
dask[complete]>=2023.9.0
ray[default]>=2.7.0
joblib>=1.3.0

# ============================================================================
# Backtesting & Trading
# ============================================================================
backtrader>=1.9.78.123
zipline-reloaded>=3.0.0
vectorbt>=0.25.0

# ============================================================================
# Utilities
# ============================================================================
# Infrastructure
loguru>=0.7.0
pyyaml>=6.0
python-dotenv>=1.0.0

# API & Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
redis>=5.0.0  # For distributed event bus

# Real-Time Data Connectors (Optional)
# alpaca-trade-api>=3.0.0  # Uncomment for Alpaca
# polygon-api-client>=1.12.0  # Uncomment for Polygon
# ib_insync>=0.9.86  # Uncomment for Interactive Brokers

# Production Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# python-multipart - already listed above
cryptography>=41.0.0

# Production Monitoring
prometheus-client>=0.18.0
aiohttp>=3.9.0

# Production Configuration
pydantic-settings>=2.0.0
# loguru - already listed above
jsonschema>=4.19.0
pydantic>=2.4.0

# ============================================================================
# Testing & Quality
# ============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0
ruff>=0.4.0
mypy>=1.5.0
pandera>=0.18.0

# ============================================================================
# Documentation
# ============================================================================
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# ============================================================================
# Deployment & Infrastructure
# ============================================================================
docker>=6.1.0
# fastapi - already listed above
# uvicorn - already listed above
celery>=5.3.0
gunicorn>=21.2.0

# ============================================================================
# Jupyter & Notebooks
# ============================================================================
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
notebook>=7.0.0
jupyterlab>=4.0.0

# ============================================================================
# Performance Profiling
# ============================================================================
line-profiler>=4.1.0
memory-profiler>=0.61.0
py-spy>=0.3.14

# ============================================================================
# Additional Scientific Computing
# ============================================================================
numba>=0.58.0
jax>=0.4.16
jaxlib>=0.4.16

# ============================================================================
# Alternative Data (Optional)
# ============================================================================
beautifulsoup4>=4.12.0
requests>=2.31.0
# aiohttp - already listed above
tweepy>=4.14.0  # Twitter data
praw>=7.7.0  # Reddit data

# ============================================================================
# Configuration Management
# ============================================================================
hydra-core>=1.3.0
omegaconf>=2.3.0
